Create Baptiste_Notebook_Index_And_Search.ipynb
I have created a Python notebook to try ElasticSearch. I created code to index and search the metadata. And also code to read, index and search in the transcription.

I created function to handle intersection, phrase and score queries. Idk how Serena's index structure will be. But, I assume that each transcription will have a list of words with the current time. I will customize it when I have the final index strcuture and make and a clean pythons script.

I also created function to try to ouput a n-minute clip of the podcast. However, it's kinda difficult (especially for intersection or score queries).